+++ heap.h
@@ -3,14 +3,14 @@
 #include <stdint.h>
 #include "values.h"
 
-uint64_t heap_add_int(int64_t val);
+uint64_t heap_add_int(int32_t val);
 uint64_t heap_add_float(double val);
 uint64_t heap_add_str(uint64_t val);
 uint64_t heap_add_identifer(uint64_t val);
-uint64_t heap_add_logical(int64_t val);
+uint64_t heap_add_logical(int32_t val);
 
 Data heap_get_data(uint64_t address);
-int64_t heap_get_int(uint64_t address);
+int32_t heap_get_int(uint64_t address);
 double heap_get_float(uint64_t address);
 uint64_t heap_get_str(uint64_t address);
-int64_t heap_get_logical(uint64_t address);
+int32_t heap_get_logical(uint64_t address);
